pyproject.toml
[project]
name = "researchclaw"
version = "0.2.0"
description = "Multi-role research loop agent system"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.scripts]
rclaw = "rclaw.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rclaw*"]
